+++ resources/views/bootstrap-4/includes/empty.blade.php
@@ -1,3 +1,0 @@
-<tr>
-    <td colspan="{{ collect($columns)->count() }}">@lang('laravel-livewire-tables::strings.no_results')</td>
-</tr>