+++ resources/views/bootstrap-5/components/table/row.blade.php
@@ -1,0 +1,12 @@
+@props(['url' => null])
+
+<tr
+    {{ $attributes }}
+
+    @if ($url)
+        onclick="window.location='{{ $url }}';"
+        style="cursor:pointer"
+    @endif
+>
+    {{ $slot }}
+</tr>